requirements.txt
openpyxl>=3.0.0
xlrd>=2.0.0
beautifulsoup4>=4.10.0
tencentcloud-sdk-python>=3.0.0
Flask>=2.0.0
python-dotenv
google-genai